+++ Source/UE5_PORTFOLIO_SEKIRO/Global/Data/PlayerStatData.h
@@ -17,10 +17,10 @@ struct UE5_PORTFOLIO_SEKIRO_API FPlayerStatData : public FTableRowBase
 	~FPlayerStatData() {}
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Data")
-	int MaxHP;
+	float MaxHP;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Data")
-	int MaxPosture;
+	float MaxPosture;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Data")
 	float Power;
